bench: add flipped join batching benchmark
